aggiunta classe CheckingAccount.java in sostituzione alla classe BankAccount.java
